Run unit tests inside python docker image
Accepting the reality that there are two codebases here, with two
different python versions/dependencies, split the test suite into two.

The functional tests test the rendering and actual execution of the
analysis code.

The unit tests test the specifics of the code itself, and run inside the
python docker image.