Further tweaks running the tests in Docker
To fix the relative imports issue, we switch to running the script via
`python -m analysis.measures`. This allows us to use an absolute import
from within measures.py, which works both when run as an action, and
when importing for testing.

Also improved test-functional to work with running each templated
analysis in isolation. The previous sys.path munging had the potential
to break confusingly when we add a new templates analysis, so this is
much cleaner