Convert into a python package.
This change does two main things:

1) moves the current rendering and schema code from job-server to this
   project, to faciliate better development of analysis code.

2) converts the repo to a python package that will be versioned and and
   added as a dependency to job-server

The main goal is easier development and testing of analysis code by
keeping the schema and the templates in the same repo.  It also prepares
the way for use in a future secure-service.

Developing analysis code is now a case of cd'ing into the directory, and
runing `just test`