chore: better local development for analysis code
This change makes `just render` command work when run from a template
subdirectory, e.g `interactive_templates/templates/v2`

When invoked from such a directory, we use that directory as the
template dir *and* the output dir, but we skip copying files, just
render templates.  This allows a developer to run `just render` and then
develop using `opensafely run` like normal. They'd only need to run
`just render` again if they change a templated file.

It uses the default test data for the analysis by default, but custom
template data can be override with `just render foo=bar` etc.

Doing local development in this way means that there are various local
directories that should be ignored, in `.gitingore`, `MANIFEST.in`, and
when rendering code. To make this tractable, this changes the paths of
the downloaded costlists to be in a separate subdir, so that these
various tools can exclude them.