Temporarily provide default week of latest extract
Job Server hasn't yet been updated to calculate this date, so we are providing
a default date, which is correct for now, to allow us to merge this change
without breaking anything. It is expected that Job Server will be updated
within the next day or two, at which point this commit should be reverted.