Render json config and use it instead of passing cli args
Serialising the config into cli args and parsing it was source of bugs.

Instead, we render the *validated* form input as json, and then load
that to access the data in the code.

Everything is validated and typed correctly, no more hand parsing
needed. And simplifies the project.yaml to be more readable.

We do re-use the study definition in the weekly action, so make those
still overloadable via cohortextractor parameters.