Add `exceptiongroup` to test requirements
To ensure it's not removed by Dependabot.

This is a workaround for working with Dependabot and multiple Python
versions. It is a minimally disruptive way of ensuring that Dependabot
doesn't remove this dependency.

When checking these test requirements with Python 3.11, it is fine to
remove `exceptiongroup` because `hypothesis` only uses it for Python <
3.11. However, the tests run with Python 3.8.

We're currently limited with Python 3.8 due to the use of the OpenSAFELY
`python:v1` image. Even if we upgrade to `python:v2`, that has Python
3.10

Dependabot runs with a specific Python version for a repository. The
code is designed to run with Python 3.11. If we configured it to run
with Python 3.8; it may break requirements in the actual code
dependencies.

Instead go with explicitly specifying `exceptiongroup` as a requirement,
which was @iaindillingham's suggestion in #379.